+++ code/sspp-assignment/common/ELLPACKTransformer.h
@@ -16,17 +16,17 @@ namespace sspp {
       template<typename VALUE_TYPE, typename INPUT_TYPE>
       static ELLPACK<VALUE_TYPE> transform(MatrixMarket<INPUT_TYPE> & matrix_market) {
         std::vector<MatrixMarketTuple<INPUT_TYPE>> tuples = matrix_market.GetTuples();
-        std::vector<unsigned> auxiliary_vector(matrix_market.GetRows());
-        for(unsigned i = 0; i < matrix_market.GetNonZeros(); i++) {
+        std::vector<unsigned long long> auxiliary_vector(matrix_market.GetRows());
+        for(unsigned long long i = 0; i < matrix_market.GetNonZeros(); i++) {
           ++auxiliary_vector[tuples[i].GetRowIndice()];
         }
 
-        unsigned max_row_non_zeros = *std::max_element(auxiliary_vector.begin(), auxiliary_vector.end());
-        std::vector<unsigned> column_indices(max_row_non_zeros * matrix_market.GetRows());
+        unsigned long long max_row_non_zeros = *std::max_element(auxiliary_vector.begin(), auxiliary_vector.end());
+        std::vector<unsigned long long> column_indices(max_row_non_zeros * matrix_market.GetRows());
         std::vector<VALUE_TYPE> values(max_row_non_zeros * matrix_market.GetRows());
 
-        for(unsigned i = 0; i < matrix_market.GetNonZeros(); ++i) {
-          unsigned index = max_row_non_zeros * tuples[i].GetRowIndice() + auxiliary_vector[tuples[i].GetRowIndice()] - 1;
+        for(unsigned long long i = 0; i < matrix_market.GetNonZeros(); ++i) {
+          unsigned long long index = max_row_non_zeros * tuples[i].GetRowIndice() + auxiliary_vector[tuples[i].GetRowIndice()] - 1;
           column_indices[index] = tuples[i].GetColumnIndice();
           values[index] = static_cast<VALUE_TYPE>(tuples[i].GetValue());
           --auxiliary_vector[tuples[i].GetRowIndice()];
