+++ code/sspp-assignment/common/SparseMatrix.h
@@ -5,26 +5,26 @@ namespace sspp {
   namespace common {
     class SparseMatrix {
     public:
-      SparseMatrix(unsigned rows, unsigned columns, unsigned non_zeros)
+      SparseMatrix(unsigned long long rows, unsigned long long columns, unsigned long long non_zeros)
         : rows_(rows), columns_(columns), non_zeros_(non_zeros) {
       };
 
-      unsigned GetNonZeros() const {
+      unsigned long long GetNonZeros() const {
         return non_zeros_;
       };
 
-      unsigned GetRows() const {
+      unsigned long long GetRows() const {
         return rows_;
       }
 
-      unsigned GetColumns() const {
+      unsigned long long GetColumns() const {
         return columns_;
       };
 
     protected:
-      unsigned rows_;
-      unsigned columns_;
-      unsigned non_zeros_;
+      unsigned long long rows_;
+      unsigned long long columns_;
+      unsigned long long non_zeros_;
     };
   }
 }
